Test(tutorial5) : Just tested some additions with different types

+++ tutorial5/src/main.rs
@@ -1,7 +1,7 @@
 fn main() {
-    let x: u8 = 9; // 0 <-> 255
+    let x: i64 = 9; // 0 <-> 255
     let y: i8 = 10; // -128 <-> 127
 
     let z = x + y;
-    println!("{}", z); // cannot do this because no implementation for `u8 + i8`
+    println!("{}", z); // cannot do this because no implementation for `i64 + i8`
 }
